requirements.txt
streamlit
pycryptodome